+++ app/Views/layout/sidebar.php
@@ -84,6 +84,18 @@
           </a>
         </li>
 
+        <li class="nav-item mt-3">
+          <a href="<?= base_url('logout') ?>" class="nav-link bg-danger">
+
+            <i class="nav-icon fas fa-sign-out-alt"></i>
+
+            <p>
+              Logout
+            </p>
+
+          </a>
+        </li>
+
       </ul>
     </nav>